requirements.txt
pandas
tabulate
flask
google-genai
python-dotenv